+++ src/mfx/Model.h
@@ -44,6 +44,8 @@ http://sam.zoy.org/wtfpl/COPYING for more details.
 #include <array>
 #include <map>
 
+#include <cstdint>
+
 
 
 namespace mfx
@@ -175,7 +177,7 @@ public:
 	int            start_d2d_rec (const char pathname_0 [], size_t max_len);
 	int            stop_d2d_rec ();
 	bool           is_d2d_recording () const;
-	size_t         get_d2d_size_frames () const;
+	int64_t        get_d2d_size_frames () const;
 
 	static const std::array <int, Cst::_nbr_pedals> // [Pedal number] = Input switch index
 	               _pedal_to_switch_map;
