+++ src/mfx/RecD2d.cpp
@@ -171,7 +171,7 @@ bool	RecD2d::is_recording () const
 
 
 
-size_t	RecD2d::get_size_frames () const
+int64_t	RecD2d::get_size_frames () const
 {
 	return _file_writer.get_size_frames ();
 }
