+++ src/mfx/RecD2d.h
@@ -35,6 +35,7 @@ http://www.wtfpl.net/ for more details.
 #include <vector>
 
 #include <cstdio>
+#include <cstdint>
 
 
 
@@ -57,7 +58,7 @@ public:
 	int            close_file ();
 	int            process_messages ();
 	bool           is_recording () const;
-	size_t         get_size_frames () const;
+	int64_t        get_size_frames () const;
 
 	// RT thread
 	void           write_data (const float * const chn_arr [], int nbr_spl);
